+++ internal/client/persistent.go
@@ -66,7 +66,8 @@ type Persistent struct {
 
 	SafeSearch filtering.SafeSearch
 
-	// BlockedServices is the configuration of blocked services of a client.
+	// BlockedServices is the configuration of blocked services of a client.  It
+	// must not be nil after initialization.
 	BlockedServices *filtering.BlockedServices
 
 	Name string